Mayor restructuring in order to easy support the creation of multiple shaders. Decreasing reliability to SharpGL.SceneGraph.dll. Add restrains to force user into best practices. Virtual Mesh builder with auto normal calculation. Attempted triangulation. Easy to find the use of particle systems. Better, faster and more dynamic viewport than the SharpGL.SceneGraph one.

+++ SharpGLHelper/SharpGLHelper/IO/IFileModel.cs
@@ -18,7 +18,7 @@ namespace SharpGLHelper.IO
         /// <summary>
         /// The indices.
         /// </summary>
-        float[] Indices { get; set; }
+        uint[] Indices { get; set; }
         /// <summary>
         /// The normals.
         /// </summary>
